UAS PBO 2011522012RAHMADINA

+++ src/Bbm.java
@@ -1,11 +1,12 @@
 import java.sql.SQLException;
 
+//interface bbm
 public interface Bbm {
     public int noFaktur(int iharga) throws SQLException;
     public String admin(String iadm);
     public int harga(int iharga);
     public String tanggal(int iharga);
-    public double jualPremium(int iharga);
-    public double jualPertalite(int iharga);
-    public double jualPertamax(int iharga);
+    public float jualPremium(int iharga);
+    public float jualPertalite(int iharga);
+    public float jualPertamax(int iharga);
 }
